2. Тесты для идеальный вес [#325686]

+++ src/ru/job4j/calculator/Fit.java
@@ -1,12 +1,12 @@
 package ru.job4j.calculator;
 
 public class Fit {
-    public static double manWeight(short height) {
+    public static double manWeight(double height) {
         double rsl = (height - 100) * 1.15;
         return rsl;
     }
 
-    public static double womanWeight(short height) {
+    public static double womanWeight(double height) {
         double rsl = (height - 110) * 1.15;
         return rsl;
     }
